requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain-huggingface
faiss-cpu
langchain-chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube-transcript-api
unstructured
pytube
numexpr
huggingface-hub
streamlit
streamlit-chat
uuid
pytesseract
pdf2image
Pillow
pickle-mixin
trulens
shap
evaluate
ipykernel
